telegram_shop_bot/requirements.txt
python-telegram-bot[ext]
SQLAlchemy
python-dotenv